requirements.txt
# Core dependencies
neo4j==5.14.0
streamlit==1.28.0
langchain==0.1.0
langchain-community==0.0.10

# LLM providers
openai==1.6.1
huggingface-hub==0.20.0
transformers==4.36.0

# Embeddings
sentence-transformers==2.2.2

# Data processing
pandas==2.1.4
numpy==1.24.3

# NLP
spacy==3.7.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
